requirements.txt
tgtg~=0.15.0

python-telegram-bot~=20.3
